requirements_portable.txt
# requirements_py37.txt - Compatible with Python 3.7.16
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
asynctest==0.13.0
attrs==24.2.0
backcall==0.2.0
boto3==1.33.13
botocore==1.33.13
chardet==5.2.0
click==8.1.8
Cython==3.0.12
datasets==2.13.2
debugpy==1.7.0
decorator==5.1.1
dill==0.3.6
entrypoints==0.4
filelock==3.12.2
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
huggingface-hub==0.16.4
idna==3.10
imath==0.0.1
importlib-metadata==6.7.0
iopath==0.1.10
ipykernel==6.16.2
ipython==7.34.0
jedi==0.19.2
jmespath==1.0.1
joblib==1.3.2
jupyter_client==7.4.9
jupyter_core==4.12.0
kiwisolver==1.4.5
matplotlib==3.5.3
matplotlib-inline==0.1.6
# Use compatible MKL versions
mkl-fft==1.3.0
mkl-random==1.2.2  # Updated from 1.1.1 which is not available
mkl-service==2.3.0
multidict==6.0.5
multiprocess==0.70.14
munkres==1.1.4
nest-asyncio==1.6.0
nltk==3.8.1
numpy==1.21.6
# Only include CUDA packages if you need GPU support
# nvidia-cublas-cu11==11.10.3.66
# nvidia-cuda-nvrtc-cu11==11.7.99
# nvidia-cuda-runtime-cu11==11.7.99
# nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.11.0.86
packaging==24.0
pandas==1.3.5
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
portalocker==2.7.0
prompt_toolkit==3.0.48
protobuf==4.24.4
psutil==7.0.0
ptyprocess==0.7.0
py360convert==0.1.0
pyarrow==12.0.1
Pygments==2.17.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
PyYAML==6.0.1
pyzmq==26.2.1
regex==2024.4.16
requests==2.31.0
s3transfer==0.8.2
# Use scikit-learn and scipy versions compatible with Python 3.7
scikit-learn==1.0.2
scipy==1.7.3
six==1.17.0
tabulate==0.9.0
tensorboardX==2.6.2.2
termcolor==2.3.0
tornado==6.2
tqdm==4.67.1
traitlets==5.9.0
urllib3==1.26.20
wcwidth==0.2.13
xxhash==3.5.0
yacs==0.1.8
yarl==1.9.4